moved schemas back into catalog
