+++ services/catalog/src/simcore_service_catalog/models/schemas/services.py
@@ -1,0 +1,20 @@
+from typing import Optional
+
+from pydantic import EmailStr
+
+from models_library.services import (
+    ServiceAccessRights,
+    ServiceDockerData,
+    ServiceMetaData,
+)
+
+
+# OpenAPI models (contain both service metadata and access rights)
+class ServiceUpdate(ServiceMetaData, ServiceAccessRights):
+    pass
+
+
+class ServiceOut(
+    ServiceDockerData, ServiceAccessRights, ServiceMetaData
+):  # pylint: disable=too-many-ancestors
+    owner: Optional[EmailStr]
